feat: refactor file handler imports and enhance BuilderContext with setNodeData method
